Папки, квадратка, draw, double

+++ Stack/stack.cpp
@@ -28,7 +28,7 @@ static const size_t CANARY = 1910;
 
 static const char POISON     = '@';    //  '@' == 64
 static const size_t MIN_SIZE = 100;
-static const size_t MAX_SIZE = -1;     // size_t is unsigned that's why it will make maximum value of stack
+static const size_t MAX_SIZE = (size_t) -1;     // size_t is unsigned that's why it will make maximum value of stack
 static const int STACK_SCALE = 2;
 static const size_t KEY      = (size_t) time (NULL);
